+++ IDHIUtils.Core/Constants.cs
@@ -17,6 +17,15 @@ namespace IDHIUtils
     #endregion
 
     #region enums
+    public enum HeroineStatus
+    {
+        Unknown,
+        Safe,
+        Risky,
+        Pregnant,
+        OnLeave
+    }
+
     public enum PositionCategory
     {
         LieDown = 0,
@@ -31,10 +40,40 @@ namespace IDHIUtils
         StandPool = 9,
         SitDesk = 10,
         SquadDesk = 11,
-        Pool = 1004,
+        BenchBlowjob = 12,
+        Bookshelf = 1002,
         MischievousCaress = 1003,
-        Ground3P = 1100,
-        AquariumCrowded = 1304,
+        Pool = 1004,
+        Fence = 1006,
+        PiledriverMissionary = 1008,
+        BoxDoggy = 1009,
+        CornerMasturbation = 1010,
+        SeatedMasturbation = 1011,
+        StandingMasturbation = 1012,
+        BedMasturbation = 1014,
+        TennisNetMasturbation = 1015,
+        Dildo = 1016,
+        Rotor = 1017,
+        LesbianLieDown = 1100,
+        LesbianChair = 1101,
+        LesbianStand = 1102,
+        Groping = 1102,
+        StraddleBench = 1200,
+        SofaCowgirl = 1201,
+        TennisNet = 1300,
+        TableTennis = 1301,
+        AgainstWall = 1302,
+        BananaBoat = 1303,
+        Crowded = 1304,
+        DeskSquatting = 1305,
+        Tub = 1306,
+        PoolFloats = 1307,
+        ToiletPeeping = 2000,
+        ShowerPeeping = 2001,
+        ToiletMasturbation = 2003,
+        ShowerMasturbation = 2004,
+        LieDown3P = 3000,
+        Sitting3P = 3001
     }
 
     public enum ErogenousZone